Inline logging and serialization helpers
