+++ quality_assurance/main.py
@@ -10,6 +10,7 @@ import os
 import sys
 import asyncio
 import logging
+from logging.handlers import RotatingFileHandler
 import subprocess
 import re
 from typing import Dict, List, Optional, Any, Union
@@ -55,7 +56,7 @@ except AttributeError:  # Pillow < 9.1
 from prometheus_client import Counter, Histogram, Gauge, start_http_server, generate_latest, REGISTRY
 from prometheus_client.exposition import CONTENT_TYPE_LATEST
 import psutil
-from logging_utils import configure_logging
+import structlog
 
 # Наши валидаторы
 from ocr_validator import OCRValidator, OCRValidationConfig
@@ -105,6 +106,82 @@ class Settings(BaseSettings):
         env_file = ".env"
 
 settings = Settings()
+
+
+DEFAULT_MAX_BYTES = 10 * 1024 * 1024
+DEFAULT_BACKUP_COUNT = 5
+
+
+def configure_logging(
+    component_group: str,
+    component_name: str,
+    *,
+    debug: bool = False,
+    log_directory: Optional[Path] = None,
+):
+    level = logging.DEBUG if debug else logging.INFO
+
+    log_dir = Path(log_directory) if log_directory is not None else Path("logs")
+    log_dir.mkdir(parents=True, exist_ok=True)
+    log_file = log_dir / f"{component_group}.log"
+
+    timestamper = structlog.processors.TimeStamper(fmt="iso")
+
+    structlog.configure(
+        processors=[
+            structlog.stdlib.add_logger_name,
+            structlog.stdlib.add_log_level,
+            timestamper,
+            structlog.stdlib.ProcessorFormatter.wrap_for_formatter,
+        ],
+        context_class=dict,
+        logger_factory=structlog.stdlib.LoggerFactory(),
+        wrapper_class=structlog.stdlib.BoundLogger,
+        cache_logger_on_first_use=True,
+    )
+
+    formatter = structlog.stdlib.ProcessorFormatter(
+        foreign_pre_chain=[
+            structlog.stdlib.add_logger_name,
+            structlog.stdlib.add_log_level,
+            timestamper,
+        ],
+        processor=(
+            structlog.dev.ConsoleRenderer()
+            if debug
+            else structlog.processors.JSONRenderer(sort_keys=True)
+        ),
+    )
+
+    file_handler = RotatingFileHandler(
+        log_file,
+        maxBytes=DEFAULT_MAX_BYTES,
+        backupCount=DEFAULT_BACKUP_COUNT,
+        encoding="utf-8",
+    )
+    file_handler.setFormatter(formatter)
+    file_handler.setLevel(level)
+
+    stream_handler = logging.StreamHandler()
+    stream_handler.setFormatter(formatter)
+    stream_handler.setLevel(level)
+
+    root_logger = logging.getLogger()
+    root_logger.setLevel(level)
+    for handler in list(root_logger.handlers):
+        root_logger.removeHandler(handler)
+    root_logger.addHandler(file_handler)
+    root_logger.addHandler(stream_handler)
+
+    logger = structlog.get_logger(component_name)
+    logger.info(
+        "logging configured",
+        log_group=component_group,
+        log_file=str(log_file),
+        debug=debug,
+    )
+    return logger
+
 
 # Настройка логирования объединена для документного конвейера
 logger = configure_logging(
